Major World-Gen Overhaul: Continental Plates, Erosion, Hydrology, and Spawn Selection
- Globe Generation Pipeline: Implement a 6-stage physical simulation:
  1. Tectonic Plates (plates.rs): Lloyd-relaxed Voronoi plates with motion vectors and uplift/subsidence.
  2. Multi-Octave Heightmap: Blended noise and plate-based mountain ridges.
  3. Erosion (erosion.rs): Thermal slumping and sediment-carrying hydraulic erosion.
  4. Hydrology (hydrology.rs): Priority-flood pit filling, D8 flow accumulation, and river/lake extraction.
  5. Climate (climate.rs): Latitudinal temperature and orographic rainfall modeling.
  6. Biome (biome.rs): Whittaker classification with bilinear sampling for continuous transitions.
- Chunk Generation: Rivers and lakes are now deterministically stamped into local chunks during streaming.
- Game Lifecycle:
  - Add GameState::SpawnSelect to allow players to inspect the globe and choose a starting region.
  - Decouple globe cells (climate) from mega-chunks (settlement regions).
- Persistence: Versioned world.bin cache (v2) for robust serialization of the complex globe data.
- Documentation: Updated CLAUDE.md with technical mandates for the new world pipeline.

+++ src/pathfinding/chunk_router.rs
@@ -47,7 +47,7 @@ impl ChunkRouter {
         cur: ChunkCoord,
         dest: ChunkCoord,
         current_z: i8,
-    ) -> Option<(i16, i16)> {
+    ) -> Option<(i32, i32)> {
         if cur == dest {
             return None;
         }
@@ -109,7 +109,7 @@ impl ChunkRouter {
         let chosen = best?;
         let gx = chosen.neighbor.0 * CHUNK_SIZE as i32 + chosen.entry_local.0 as i32;
         let gy = chosen.neighbor.1 * CHUNK_SIZE as i32 + chosen.entry_local.1 as i32;
-        Some((gx as i16, gy as i16))
+        Some((gx as i32, gy as i32))
     }
 
     pub fn cached_destination_count(&self) -> usize {
@@ -261,6 +261,6 @@ mod tests {
         let wp = router
             .first_waypoint(&graph, ChunkCoord(0, 0), ChunkCoord(2, 0), 0)
             .unwrap();
-        assert_eq!(wp.0, CHUNK_SIZE as i16);
+        assert_eq!(wp.0, CHUNK_SIZE as i32);
     }
 }
